[bug/12] Rename everything back
