+++ tests/decode-json.test.ts
@@ -1,6 +1,6 @@
 import test from 'ava'
 
-import Decode from '../src'
+import Decode from '../src/decode-json'
 import { InvalidJson, InField, RuntimeException } from './error'
 
 test('Decoder.decodeJSON()', t => {
