+++ tests/record.test.ts
@@ -2,7 +2,7 @@
 
 import test from 'ava'
 
-import Decode from '../src'
+import Decode from '../src/decode-json'
 import {
   Optional,
   InField,
